pyproject.toml
[tool.poetry]
name = "standup"
version = "0.1.0"
description = ""
authors = ["skippi <jbtcao@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
invoke = "^1.2"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
pylint = "^2.3"

[tool.poetry.scripts]
standup = 'standup.shell:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
